blog-jinja2-auth/pyproject.toml
[tool.poetry]
name = "bareASGI-blog"
version = "0.1.0"
description = ""
authors = ["Rob Blackbourn <rob.blackbourn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aiosqlite = "^0.10.0"
bareASGI = "^3.3"
bareasgi-jinja2 = "^3.0"
hypercorn = "^0.9.0"
PyYaml = "^5.1"
PyJWT = "^1.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4"
mypy = "^0.740.0"
autopep8 = "^1.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
